location refactor
