+++ server/helpers/api.js
@@ -2,8 +2,8 @@ module.exports = version => {
   return {
     users: `/api/${version}/users`,
     usersRoute: `./routes/api/${version}/users`,
-    location: `/api/${version}/locations`,
-    locationsRoute: `./routes/api/${version}/locations`,
+    location: `/api/${version}/location`,
+    locationsRoute: `./routes/api/${version}/location`,
     reviews: `/api/${version}/reviews`,
     reviewsRoute: `./routes/api/${version}/reviews`,
     createCustomerWithLocation: `/api/${version}/create/location-with-customer`,
