+++ server/routes/api/v1/location.js
@@ -1,15 +1,15 @@
-const locationController = require('../../../controllers/locations');
+const locationController = require('../../../controllers/location');
 const router = require('express').Router();
 
   router
     .route('/')
-    .get(locationController.getLocations)
+    .get(locationController.getAll)
     .post(locationController.addLocation)
     .put(() => false)
     .delete(() => false);
 
   router
-    .route('/:homeId').get(locationController.getLocation);
+    .route('/:id').get(locationController.getLocation);
 
     module.exports = router;
